+++ CHANGELOG.md
@@ -5,6 +5,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Breaking Changes
+- Changed AzureAD parameters to EntraID
+
+### Fixed
+- Fixed an issue with the get requests when the return object does not have a value
+
+### Added
+- Added new functions for Intune Policies
+    - Get-IntuneDriverUpdatePolicy
+    - Get-IntuneDriverUpdatePolicyAssignments
+    - New-IntuneDriverUpdatePolicyAssignment
+    - New-IntuneDriverUpdatePolicyAssignment
+
 ## [1.0.21] - 2023-05-10
 
 ### Fixed